Fix logic for plotting arrival of different phases
When deciding whether to plot a particular arrival time, use the event
time and the time the graph has started.  Before this, we'd get a
graph that might include (say) the time of arrival of the P-wave, but
the calculations in this function would skip it on the assumption that
we weren't graphing it.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>